requirements.txt
# Core ML and Data Science
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Arabic NLP Specific
arabert>=0.1.0
pyarabic>=0.6.14
tashkeel>=0.0.1

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
argparse>=1.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Arabic Text Processing
regex>=2023.6.0
emoji>=2.8.0

# Model Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2

# Optional: GPU Support
# nvidia-ml-py3>=7.352.0
